Use handoffMap for switching between content and fallback
Previously when using switchMap there was an issue with acquiring a
screenshare because the fallbackTrack would be unsubscribed from
immediately, while the screenshare track could take as long as the user
wants to make their selection. This resulted in frames in the fallback
track stopping, and could prevent the track from being pulled
successfully in some cases.